requirements.txt
Flask==3.1.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==20.1.0
psutil==5.9.5
setuptools==79.0.1
wheel==0.40.0